v0.4 Scale, Intraday data improvements, and enums
This update adds lots of improvement and a few type-changes. We're going
to use enums where possible. DevceType is one example. Intraday adds
"Level" which comes in on calories. Authenticator class adds another
method, GetAuthUrlTokenForcePromptingLogin which can give a URL that
forces the user to re-login if they are logged in with another account.