Added error handling on "access/token".
Yesterday the FitBit guys released a Security Fix for a vulnerability
they found on January 17th  which caused that the FitBit API Client
ceased to work as it was not implementing the OAuth Authentication
properly. It took a little bit of time to figure it out but I could
eventually get a fix for it.

FitBit Security Vulnerability announce:

https://groups.google.com/forum/?hl=en&lnk=gcimh#!topic/fitbit-api/ii4pUt4uTNM