Fixes #63 - provides ability to intercept raw http request/response
I introduce an interface to allow client consumers to intercept
requests/responses. The interface is an abstraction over the
HttpMessageHandler capabilities of HttpClient, but simplifies usage for
consumers of the library. Missing: a constructor where consumer can
directly provide the implementation of their interface (instead of the
handler, like it is in current implementation)