Adding a website for debug purposes
The idea: the sample website has a reference on public nuget package. Is
stable, easy for people to get started.

This new debug website has a source dependency to Fitbit solution. It is
used for live debugging purposes (preference is to do development on
Fitbit.sln) and stays current with latest changes. Not as stable as the
sample.

Not a big fan or having to add all these files (I would put that on how
heavy MVC is. Looking forward to Core), but I think it is the best
compromise.