Add a sample for how to use the FitBit client from a console app.
The sample handles the authorization to FitBit and downloads the weight for
the current day.

Once the app is authorized, the token data is stored locally to avoid having to
re-authenticate every time. For brevity, the token is stored in plain text.

In order for the authorization to work it requires the app to be launched as
admin. That is required in order for the app to be able to listen to the
callback at 'http://localhost'. The app automatically handles this case.